requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
requests==2.32.3
anthropic==0.43.0
PyPDF2==3.0.1
python-dotenv==1.0.1
